Integer overflow when calculating standard deviation and requests take longer than 46340 ms

+++ gatling-charts/src/main/scala/io/gatling/charts/result/reader/buffers/GeneralStatsBuffers.scala
@@ -72,7 +72,7 @@ class GeneralStatsBuffer(duration: Long) extends CountBuffer {
   override def update(time: Int): Unit = {
     super.update(time)
     digest.add(time)
-    sumOfSquares += time * time
+    sumOfSquares += time.toLong * time.toLong
     sum += time
   }
